spykingcircus and kilosort extractors
