+++ spikeinterface/__init__.py
@@ -10,9 +10,10 @@ from .extractors.mearecextractors.mearecextractors import MEArecRecordingExtract
 from .extractors.biocamrecordingextractor import BiocamRecordingExtractor
 from .extractors.hs2sortingextractor import HS2SortingExtractor
 from .extractors.klustasortingextractor import KlustaSortingExtractor
-from .extractors.kilosortsortingextractor import KilosortSortingExtractor
+from .extractors.kilosortsortingextractor import KiloSortSortingExtractor
 from .extractors.numpyextractors.numpyextractors import NumpyRecordingExtractor, NumpySortingExtractor
 from .extractors.nwbextractors.nwbextractors import NwbRecordingExtractor
 from .extractors.openephysextractors.openephysextractors import OpenEphysRecordingExtractor, OpenEphysSortingExtractor
+from .extractors.spykingcircussortingextractor.spykingcircussortingextractor import SpykingCircusSortingExtractor
 
 from .tools import loadProbeFile, saveProbeFile, read_python
